requirements.txt
asgiref==3.11.1
dj-database-url==2.3.0
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg[binary]==3.2.6
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.9.0
